+++ crates/threshold-signatures/src/frost/eddsa/sign.rs
@@ -463,8 +463,8 @@ mod test {
     };
     use crate::test_utils::{
         assert_buffer_capacity, assert_public_key_invariant, build_frost_key_packages_with_dealer,
-        expected_buffer_by_role, generate_participants, generate_participants_with_random_ids,
-        one_coordinator_output, run_keygen, run_refresh, run_reshare, MockCryptoRng,
+        check_one_coordinator_output, expected_buffer_by_role, generate_participants,
+        generate_participants_with_random_ids, run_keygen, run_refresh, run_reshare, MockCryptoRng,
     };
     use crate::{
         crypto::hash::hash,
@@ -505,7 +505,7 @@ mod test {
                     &mut rng,
                 )
                 .unwrap();
-                one_coordinator_output(data, coordinator).unwrap();
+                check_one_coordinator_output(data, coordinator).unwrap();
             }
         }
     }
@@ -533,7 +533,7 @@ mod test {
                     MockCryptoRng::seed_from_u64(rng.next_u64()),
                 )
                 .unwrap();
-                one_coordinator_output(data, coordinator).unwrap();
+                check_one_coordinator_output(data, coordinator).unwrap();
             }
         }
     }
@@ -576,7 +576,7 @@ mod test {
             },
         )
         .unwrap();
-        let signature = one_coordinator_output(result, coordinator).unwrap();
+        let signature = check_one_coordinator_output(result, coordinator).unwrap();
         insta::assert_json_snapshot!(signature);
     }
 
@@ -621,7 +621,7 @@ mod test {
             },
         )
         .unwrap();
-        let signature = one_coordinator_output(result, coordinator).unwrap();
+        let signature = check_one_coordinator_output(result, coordinator).unwrap();
         insta::assert_json_snapshot!(signature);
     }
 
@@ -647,7 +647,7 @@ mod test {
                 &mut rng,
             )
             .unwrap();
-            let signature = one_coordinator_output(data, coordinator).unwrap();
+            let signature = check_one_coordinator_output(data, coordinator).unwrap();
 
             // externally verify with the signature
             assert!(key_packages[0]
@@ -682,7 +682,7 @@ mod test {
                 MockCryptoRng::seed_from_u64(rng.next_u64()),
             )
             .unwrap();
-            let signature = one_coordinator_output(data, coordinator).unwrap();
+            let signature = check_one_coordinator_output(data, coordinator).unwrap();
 
             // externally verify with the signature
             assert!(key_packages[0]
@@ -733,7 +733,7 @@ mod test {
                 &mut rng,
             )
             .unwrap();
-            let signature = one_coordinator_output(data, coordinator).unwrap();
+            let signature = check_one_coordinator_output(data, coordinator).unwrap();
 
             // externally verify with the signature
             assert!(key_packages[0]
@@ -794,7 +794,7 @@ mod test {
                 MockCryptoRng::seed_from_u64(rng.next_u64()),
             )
             .unwrap();
-            let signature = one_coordinator_output(data, coordinator).unwrap();
+            let signature = check_one_coordinator_output(data, coordinator).unwrap();
 
             // externally verify with the signature
             assert!(key_packages[0]
@@ -856,7 +856,7 @@ mod test {
                 &mut rng,
             )
             .unwrap();
-            let signature = one_coordinator_output(data, coordinator).unwrap();
+            let signature = check_one_coordinator_output(data, coordinator).unwrap();
 
             // externally verify with the signature
             assert!(key_packages[0]
@@ -1008,7 +1008,7 @@ mod test {
                 MockCryptoRng::seed_from_u64(rng.next_u64()),
             )
             .unwrap();
-            let signature = one_coordinator_output(data, coordinator).unwrap();
+            let signature = check_one_coordinator_output(data, coordinator).unwrap();
 
             // externally verify with the signature
             assert!(key_packages[0]
